12/2/2021 - 8:22 AM look at README.md

+++ src/main/java/com/group18/capstone/controller/Food.java
@@ -28,7 +28,7 @@ public class Food {
         return ItemDesc;
     }
 
-    public double getItemPrice() {
+    public float getItemPrice() {
         return ItemPrice;
     }
